requirements.txt
absl-py==0.8.1
aiohttp==3.6.2
appdirs==1.4.3
appnope==0.1.0
astor==0.8.0
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.0
black==19.10b0
bleach==3.1.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
click==6.6
cmake==3.15.3
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.7
dlib==19.18.0
emoji==0.5.4
face-recognition-models==0.3.0
gast==0.2.2
google-auth==1.7.1
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
gunicorn==19.6.0
h5py==2.10.0
html5lib==1.0.1
idna==2.8
importlib-metadata==0.23
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==0.24
jedi==0.15.1
joblib==0.14.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==0.23
matplotlib==3.1.1
mechanize==0.4.3
mistune==0.8.4
more-itertools==7.2.0
multidict==4.5.2
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
npm==0.1.1
numpy==1.17.3
oauthlib==3.1.0
opencv-python==4.1.1.26
opt-einsum==3.1.0
optional-django==0.1.0
pandas==0.25.3
pandocfilters==1.4.2
parso==0.5.1
pathspec==0.6.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
praw==6.4.0
prawcore==1.0.1
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.10.0
psaw==0.0.7
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.5
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.1
qtconsole==4.6.0
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.3.2
selenium==3.141.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
slack==0.0.2
slackclient==2.3.1
snap==0.5
soupsieve==1.9.4
sqlparse==0.3.0
tensorboard==2.0.1
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
tornado==6.0.3
traitlets==4.3.3
typed-ast==1.4.0
tzlocal==2.0.0
update-checker==0.16
urllib3==1.25.6
virtualenv==16.7.7
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
yarl==1.3.0
zipp==0.6.0
